refactor(editor): move the slash-menu view into extensions/, breaking a cycle
The CommandMenu UI component lived in the editor's shared components.tsx, so
extensions/commandMenu.ts imported it back — the edge that closed a circular
dependency (components.tsx -> extensions/index -> extensions/commandMenu ->
components.tsx).

The menu view isn't a shared shell component; it's an implementation detail of
the slash-command extension (typed by CommandItem, rendered only by that
extension's Suggestion plugin). Move it beside its extension in a new
extensions/components.tsx. commandMenu.ts now imports the sibling, so no
extensions/ file imports the editor shell and the cycle is gone (verified with
madge). commandMenu.ts stays JSX-free.
