+++ frontend/src/lib/editor/components.tsx
@@ -4,10 +4,6 @@
  * CodeMirror/OT-era editor, deleted once this cutover lands). */
 import { posToDOMRect, type Editor } from "@tiptap/core";
 import { EditorContent, useEditor } from "@tiptap/react";
-import type {
-  SuggestionKeyDownProps,
-  SuggestionProps,
-} from "@tiptap/suggestion";
 import {
   useEffect,
   useImperativeHandle,
@@ -18,9 +14,7 @@ import {
 } from "react";
 import { Awareness } from "y-protocols/awareness";
 import * as Y from "yjs";
-import { LineItemButton } from "@onyx-ai/opal/components";
 import { tiptapExtensions } from "@/lib/editor/extensions";
-import type { CommandItem } from "@/lib/editor/extensions/types";
 import {
   commentHighlights as commentHighlightPlugin,
   sourceHighlights as sourceHighlightPlugin,
@@ -461,72 +455,6 @@ export function CoeditPresenceBar({
                 : "viewing"}
           </span>
         </span>
-      ))}
-    </div>
-  );
-}
-
-export interface CommandMenuHandle {
-  onKeyDown: (props: SuggestionKeyDownProps) => boolean;
-}
-export type CommandMenuProps = {
-  ref?: Ref<CommandMenuHandle>;
-} & SuggestionProps<CommandItem>;
-export function CommandMenu({ ref, items, command }: CommandMenuProps) {
-  const [selected, setSelected] = useState(0);
-  useEffect(() => setSelected(0), [items]);
-
-  const select = (index: number) => {
-    const item = items[index];
-    if (item) command(item);
-  };
-
-  useImperativeHandle(ref, () => ({
-    onKeyDown: ({ event }) => {
-      if (event.key === "ArrowDown") {
-        setSelected((s) => (items.length ? (s + 1) % items.length : 0));
-        return true;
-      }
-      if (event.key === "ArrowUp") {
-        setSelected((s) =>
-          items.length ? (s - 1 + items.length) % items.length : 0,
-        );
-        return true;
-      }
-      if (event.key === "Enter") {
-        select(selected);
-        return true;
-      }
-      return false;
-    },
-  }));
-
-  if (items.length === 0) return null;
-
-  return (
-    <div className="max-h-[320px] w-[220px] overflow-y-auto rounded-(--radius-08) border border-(--border-01) bg-(--background-tint-00) p-1 shadow-[0px_2px_6px_var(--shadow-02),0px_0px_2px_var(--shadow-01)]">
-      {items.map((item, i) => (
-        // The wrapper carries the keyboard-selection highlight rather than
-        // relying on LineItemButton's own "selected" state, which is a hover-
-        // weight tint — too quiet for the thing Enter is about to apply. A
-        // solid fill and nothing else, as Notion's menu does it.
-        <div
-          key={item.title}
-          className={
-            i === selected
-              ? "rounded-(--radius-06) bg-(--background-tint-03)"
-              : undefined
-          }
-        >
-          <LineItemButton
-            title={item.title}
-            icon={item.icon}
-            sizePreset="main-ui"
-            variant="section"
-            state={i === selected ? "selected" : "empty"}
-            onClick={() => select(i)}
-          />
-        </div>
       ))}
     </div>
   );
